[ADD] Colors config & [UPDATE] file structure
Added the colors configuration. We can now customise the colors used by
the app as we wish by changing the hex code into the colors.json file.
Note that it doesn't update the text_colors globally. It's possible that the text is still in black even if the
color is very dark.

The already existing components have been checked to work with the new
colors config